contracts/garden: maybe-ok support for both `plant(...)` variants.
Appends data provided in call data to CREATE (deployment) code of
the seed if `plant(address,bytes)` was used.

If `plant(address)` was used, no trailing zeroes should be added
in-memory, nor subsequently copied, as the contents of
*memloc-data-size* should have stayed zero (it's only set in the
other variant).

If the above turns out to be untrue, perhaps *memloc-data-size*
can be explicitly set to 0 at some point.

Also moved *memloc-seed-code* to be last, so it doesn't overwrite
other args.