cannery: fix "no wrapper placed in memory" bug.
Was ANDing wrapper and revert opcode - no wonder! Should be ORing.

Further tests (not yet visible here) show that a transaction to
the can fails, which raises a TransactionFailed exception.

Not sure if this is expected (since the contract contains a REVERT,
after all), or if this is an indication of another failure.