cannery: 4-byte revert guard (with arg push).
To prevent exceptional halting, REVERT must be given its arguments,
so that there's no stack underflow.

Sadly, this doesn't seem to help here. Running the test still shows
the transaction's gas fully consumed.

It's likely that this is caused by the test framework not knowing
of REVERT and its reduced gas use (compared to INVALID).